Move screenshot implementation to new file
