+++ lib/sdl.cpp
@@ -724,7 +724,7 @@ void SdlWindow::mainLoop()
       }
       if (takeScreenshot)
       {
-         Screenshot(screenshot_file.c_str());
+         screenshotHelper();
          takeScreenshot = false;
       }
    }
